Basic barebones bot (Ping, Server, User). Needs all original TimeSince commands added in still
Bot currently only had discord.js's tutorials slash commands of Ping, server and user.
index.js brought down from 222 lines to 60, and should remain that way thanks to the automatic building of commands with Collection in client.commands property attached to the client instance so is a bit more readable.